Move partial message retry behavior into MessageBuffer
